backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# AI Services
openai>=1.12.0
google-genai>=1.0.0

# Data validation
pydantic>=2.6.0

# Streaming
sse-starlette>=2.0.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
